05_object_detection_yolo/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo-object-detection-api"
version = "0.1.0"
description = "Production-ready YOLO object detection API with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "OpenAI Assistant" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.4.0",
  "ultralytics>=8.3.0",
  "opencv-python-headless>=4.10.0.84",
  "numpy>=1.26.0",
  "Pillow>=10.4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
